Move Transition tests into a separate file.
